Python API: move CacheDict to common/utils.py
